Changed brooms to add fetch and delete favorites and trash

+++ src/store/order/order.controller.ts
@@ -1,11 +1,11 @@
-import { Body, Controller, Delete, Get, Param, Post, Query, Req, UseGuards, UsePipes } from '@nestjs/common';
+import { Body, Controller, Get, Param, Patch, Post, Query, Req, UseGuards, UsePipes } from '@nestjs/common';
 import { ApiHeaders, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { ObjectId } from 'mongoose';
 import { JwtAuthGuard } from 'src/auth/guard/jwt-auth.guard';
 import { ValidatePipe } from 'src/classValidator';
 import { Orders } from 'src/db-schemas/orders.schema';
 import { IReqUser } from 'src/type';
-import { BasketQueryParams, CreateOrderDto } from './dto';
+import { BasketQueryParams, CreateOrderDto, EActionBasket } from './dto';
 import { Basket } from './schema-swagger';
 import { OrderBasketService } from './services/order-basket.service';
 import { OrderService } from './services/order.service';
@@ -34,7 +34,7 @@ export class OrderController {
     return this.orderService.createOrder(createOrderDto, req.user._id);
   }
 
-  @ApiOperation({ summary: 'Add product to basket' })
+  @ApiOperation({ summary: 'Add or delete product to basket' })
   @ApiHeaders([
     {
       name: 'Authorization',
@@ -48,36 +48,17 @@ export class OrderController {
   @ApiResponse({ status: 500, description: 'Server error' })
   @UsePipes(ValidatePipe)
   @UseGuards(JwtAuthGuard)
-  @Get('add-basket/:productId')
-  addToBasket(
+  @Patch('basket/:productId')
+  updateBasketItems(
     @Param('productId') productId: ObjectId,
     @Req() req: IReqUser,
     @Query() basketQueryParams: BasketQueryParams,
   ) {
+    const { action } = basketQueryParams;
+    if (action === EActionBasket.DELETE) {
+      return this.orderBasketService.deleteToBasket(req.user._id, productId, basketQueryParams);
+    }
     return this.orderBasketService.addToBasket(req.user._id, productId, basketQueryParams);
-  }
-
-  @ApiOperation({ summary: 'Delete product from basket' })
-  @ApiHeaders([
-    {
-      name: 'Authorization',
-      required: true,
-      description: 'User access token.',
-    },
-  ])
-  @ApiResponse({ status: 200, type: Basket, description: 'Product deleted from basket' })
-  @ApiResponse({ status: 403, description: 'Invalid token' })
-  @ApiResponse({ status: 404, description: 'Invalid data' })
-  @ApiResponse({ status: 500, description: 'Server error' })
-  @UsePipes(ValidatePipe)
-  @UseGuards(JwtAuthGuard)
-  @Delete('basket/:productId')
-  deleteFromBasket(
-    @Param('productId') productId: ObjectId,
-    @Req() req: IReqUser,
-    @Query() basketQueryParams: BasketQueryParams,
-  ) {
-    return this.orderBasketService.deleteToBasket(req.user._id, productId, basketQueryParams);
   }
 
   @ApiOperation({ summary: 'Get user basket' })
@@ -93,7 +74,7 @@ export class OrderController {
   @ApiResponse({ status: 500, description: 'Server error' })
   @UsePipes(ValidatePipe)
   @UseGuards(JwtAuthGuard)
-  @Get('user-basket')
+  @Get('basket')
   getUserBasket(@Req() req: IReqUser, @Query() basketQueryParams: BasketQueryParams) {
     return this.orderBasketService.allBasketId(req.user._id, basketQueryParams);
   }
